fix: Moved emulator tests and test data to separate files
