+++ firestore-stripe-web-sdk/test/product.spec.ts
@@ -16,14 +16,7 @@
 
 import { expect, use } from "chai";
 import { fake as sinonFake, SinonSpy } from "sinon";
-import { deleteApp, FirebaseApp, initializeApp } from "@firebase/app";
-import {
-  connectFirestoreEmulator,
-  doc,
-  Firestore,
-  getFirestore,
-  setDoc,
-} from "@firebase/firestore";
+import { FirebaseApp } from "@firebase/app";
 import {
   getPrice,
   getPrices,
@@ -35,64 +28,15 @@ import {
   StripePaymentsError,
 } from "../src/index";
 import { ProductDAO, setProductDAO } from "../src/product";
+import {
+  premiumPlan,
+  premiumPlanPrice,
+  standardPlanPrice1,
+  standardPlanPrice2,
+} from "./testdata";
 
 use(require("chai-as-promised"));
-use(require("chai-like"));
 use(require("sinon-chai"));
-
-const premiumPlan: Product = {
-  active: true,
-  customProperty: "customValue",
-  description: "Access to all our content",
-  id: "premium",
-  images: [],
-  metadata: {
-    firebaseRole: "moderator",
-  },
-  name: "Premium plan",
-  prices: [],
-  role: "moderator",
-};
-
-const premiumPlanPrice: Price = {
-  active: true,
-  customProperty: "customValue",
-  currency: "usd",
-  description: "Test price 1",
-  id: "price1",
-  interval: "month",
-  intervalCount: null,
-  productId: "premium",
-  trialPeriodDays: null,
-  type: "recurring",
-  unitAmount: 999,
-};
-
-const standardPlanPrice1: Price = {
-  active: true,
-  currency: "usd",
-  description: "Test price 1",
-  id: "price1",
-  interval: "month",
-  intervalCount: null,
-  productId: "standard",
-  trialPeriodDays: null,
-  type: "recurring",
-  unitAmount: 899,
-};
-
-const standardPlanPrice2: Price = {
-  active: true,
-  currency: "usd",
-  description: "Test price 1",
-  id: "price2",
-  interval: "year",
-  intervalCount: null,
-  productId: "standard",
-  trialPeriodDays: null,
-  type: "recurring",
-  unitAmount: 9999,
-};
 
 const app: FirebaseApp = {
   name: "mock",
@@ -292,222 +236,3 @@ function testProductDAO(
 
   return (nameOrFakes as unknown) as ProductDAO;
 }
-
-/**
- * An interface for raw product data stored in Firestore. Each product is saved as a separate
- * document in the `productsCollection`. Each product can have 0 or more prices stored in
- * the `prices` subcollection.
- */
-interface ProductData {
-  product: Record<string, any>;
-  prices: Record<string, Record<string, any>>;
-}
-
-const rawTestData: Record<string, ProductData> = {
-  premium: {
-    product: {
-      active: true,
-      customProperty: "customValue",
-      description: "Access to all our content",
-      images: [],
-      metadata: {
-        firebaseRole: "moderator",
-      },
-      name: "Premium plan",
-      role: "moderator",
-    },
-    prices: {
-      price1: {
-        active: true,
-        customProperty: "customValue",
-        currency: "usd",
-        description: "Test price 1",
-        interval: "month",
-        interval_count: null,
-        trial_period_days: null,
-        type: "recurring",
-        unit_amount: 999,
-      },
-    },
-  },
-  standard: {
-    product: {
-      active: true,
-      description: "Product description",
-      images: [],
-      metadata: {},
-      name: "Product name",
-      role: null,
-    },
-    prices: {
-      price1: {
-        active: true,
-        currency: "usd",
-        description: "Test price 1",
-        interval: "month",
-        interval_count: null,
-        trial_period_days: null,
-        type: "recurring",
-        unit_amount: 899,
-      },
-      price2: {
-        active: true,
-        currency: "usd",
-        description: "Test price 1",
-        interval: "year",
-        interval_count: null,
-        trial_period_days: null,
-        type: "recurring",
-        unit_amount: 9999,
-      },
-    },
-  },
-  test: {
-    product: {
-      active: true,
-      description: "Product description",
-      images: [],
-      metadata: {},
-      name: "Product name",
-      role: null,
-    },
-    prices: {},
-  },
-};
-
-describe("Product emulator tests", () => {
-  const emulatedApp: FirebaseApp = initializeApp({
-    projectId: "fake-project-id",
-  });
-
-  const emulatedPayments: StripePayments = getStripePayments(emulatedApp, {
-    customersCollection: "customers",
-    productsCollection: "products",
-  });
-
-  before(async () => {
-    const db: Firestore = getFirestore(emulatedApp);
-    connectFirestoreEmulator(db, "localhost", 8080);
-
-    for (const [productId, data] of Object.entries(rawTestData)) {
-      await addProductData(db, productId, data);
-    }
-  });
-
-  after(async () => {
-    await deleteApp(emulatedApp);
-  });
-
-  describe("getProduct()", () => {
-    it("should return product with the specified ID", async () => {
-      const product: Product = await getProduct(emulatedPayments, "premium");
-
-      expect(product).to.eql(premiumPlan);
-    });
-
-    it("should return product with prices when includePrices is set", async () => {
-      const product: Product = await getProduct(emulatedPayments, "premium", {
-        includePrices: true,
-      });
-
-      const expected: Product = { ...premiumPlan, prices: [premiumPlanPrice] };
-      expect(product).to.be.like(expected);
-    });
-
-    it("should reject with not-found error when the specified product does not exist", async () => {
-      const err: any = await expect(
-        getProduct(emulatedPayments, "unavailable")
-      ).to.be.rejectedWith("No product found with the ID: unavailable");
-
-      expect(err).to.be.instanceOf(StripePaymentsError);
-      expect(err.code).to.equal("not-found");
-      expect(err.cause).to.be.undefined;
-    });
-  });
-
-  describe("getPrice()", () => {
-    it("should return price for the specified product and price ID", async () => {
-      const price: Price = await getPrice(
-        emulatedPayments,
-        "premium",
-        "price1"
-      );
-
-      expect(price).to.include(premiumPlanPrice);
-    });
-
-    it("should reject with not-found when the specified product does not exist", async () => {
-      const err: any = await expect(
-        getPrice(emulatedPayments, "unavailable", "price1")
-      ).to.be.rejectedWith(
-        "No price found with the product ID: unavailable and price ID: price1"
-      );
-
-      expect(err).to.be.instanceOf(StripePaymentsError);
-      expect(err.code).to.equal("not-found");
-      expect(err.cause).to.be.undefined;
-    });
-
-    it("should reject with not-found when the specified price does not exist", async () => {
-      const err: any = await expect(
-        getPrice(emulatedPayments, "premium", "unavailable")
-      ).to.be.rejectedWith(
-        "No price found with the product ID: premium and price ID: unavailable"
-      );
-
-      expect(err).to.be.instanceOf(StripePaymentsError);
-      expect(err.code).to.equal("not-found");
-      expect(err.cause).to.be.undefined;
-    });
-  });
-
-  describe("getPrices()", () => {
-    it("should return the only price as an array for the specified product ID", async () => {
-      const prices: Price[] = await getPrices(emulatedPayments, "premium");
-
-      expect(prices)
-        .to.be.an("array")
-        .of.length(1)
-        .and.be.like([premiumPlanPrice]);
-    });
-
-    it("should return prices as an array for the specified product ID", async () => {
-      const prices: Price[] = await getPrices(emulatedPayments, "standard");
-
-      expect(prices)
-        .to.be.an("array")
-        .of.length(2)
-        .and.be.like([standardPlanPrice1, standardPlanPrice2]);
-    });
-
-    it("should return empty array for existing product with no prices", async () => {
-      const prices: Price[] = await getPrices(emulatedPayments, "test");
-
-      expect(prices).to.be.an("array").and.be.empty;
-    });
-
-    it("should reject with not-found when the specified product does not exist", async () => {
-      const err: any = await expect(
-        getPrices(emulatedPayments, "unavailable")
-      ).to.be.rejectedWith("No product found with the ID: unavailable");
-
-      expect(err).to.be.instanceOf(StripePaymentsError);
-      expect(err.code).to.equal("not-found");
-      expect(err.cause).to.be.undefined;
-    });
-  });
-
-  async function addProductData(
-    db: Firestore,
-    productId: string,
-    data: ProductData
-  ): Promise<void> {
-    await setDoc(doc(db, payments.productsCollection, productId), data.product);
-    for (const [priceId, price] of Object.entries(data.prices)) {
-      await setDoc(
-        doc(db, payments.productsCollection, productId, "prices", priceId),
-        price
-      );
-    }
-  }
-});
